requirements.txt
# Game
pygame>=2.6.1

# DnD 5e Core (local package)
git+https://github.com/rapotekhin/dnd-5e-core.git

# LLM
openai>=1.12.0
python-dotenv>=1.0.0

# Data
pydantic>=2.5.3
requests>=2.31.0

FantasyNameGenerator==0.1.2

langgraph==1.0.7